fix: apply prettier fixes

+++ contracts/commands/command_swap.sol
@@ -9,23 +9,31 @@ import "../../interfaces/IERC20.sol";
 
 interface IROUTER {
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 }
 
 contract CSwap {
-
-    function swapExactTokensForTokens(uint256 amountIn, uint256 amountOutMin, address[] calldata path) external {
-        require(path.length>1,"invalid path");
+    function swapExactTokensForTokens(
+        uint256 amountIn,
+        uint256 amountOutMin,
+        address[] calldata path
+    ) external {
+        require(path.length > 1, "invalid path");
         IROUTER router = IROUTER(0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D);
         IERC20 tokenIn = IERC20(path[0]);
         tokenIn.approve(address(router), amountIn);
         // solhint-disable-next-line not-rely-on-time
-        router.swapExactTokensForTokens(amountIn, amountOutMin, path, msg.sender, block.timestamp+1);
+        router.swapExactTokensForTokens(
+            amountIn,
+            amountOutMin,
+            path,
+            msg.sender,
+            block.timestamp + 1
+        );
     }
-
 }
